build sidebar-collapse,handle function
