+++ src/components/Sidebar.tsx
@@ -1,136 +1,50 @@
-import {
-    AiOutlineHome,
-    AiFillHome,
-    AiOutlineCompass,
-    AiFillCompass,
-    AiOutlineClockCircle,
-    AiFillClockCircle,
-    AiOutlineStar,
-    AiFillStar,
-    AiOutlineSetting,
-    AiTwotoneSetting,
-} from 'react-icons/ai'
-import { HiOutlineUserGroup, HiUserGroup } from 'react-icons/hi'
-import { TfiAlarmClock } from 'react-icons/tfi'
-import { FcAlarmClock } from 'react-icons/fc'
-import { BsBookmarkStar, BsBookmarkStarFill } from 'react-icons/bs'
-import { BiDownload } from 'react-icons/bi'
-import { IoHelpCircleOutline, IoHelpCircleSharp } from 'react-icons/io5'
-import { MdMovie, MdLocalMovies } from 'react-icons/md'
-import { RiMovie2Line, RiMovie2Fill } from 'react-icons/ri'
-
-import SidebarRow from './SidebarRow'
+import { navbarData } from '@/mockapi/navbar'
+import { sidebarData1, sidebarData2, sidebarData3 } from '@/mockapi/sidebar'
 import clsx from 'clsx'
 import SidebarMenu from './SidebarMenu'
+import { useState } from 'react'
+import { AiOutlineLeft, AiOutlineRight } from 'react-icons/ai'
 
 interface Props {
     className?: string
 }
 
-const sidebarData1 = [
-    {
-        href: '/',
-        title: 'Home',
-        Icon: AiOutlineHome,
-        ActiveIcon: AiFillHome,
-        active: true,
-    },
-
-    {
-        href: 'discovery',
-        title: 'Discovery',
-        Icon: AiOutlineCompass,
-        ActiveIcon: AiFillCompass,
-    },
-    {
-        href: 'community',
-        title: 'Community',
-        Icon: HiOutlineUserGroup,
-        ActiveIcon: HiUserGroup,
-    },
-    {
-        href: 'coming soon',
-        title: 'Coming soon',
-        Icon: TfiAlarmClock,
-        ActiveIcon: FcAlarmClock,
-    },
-]
-
-const navbarData = [
-    {
-        href: '/movies',
-        title: 'Movie',
-        Icon: MdMovie,
-        ActiveIcon: MdMovie,
-    },
-    {
-        href: 'series',
-        title: 'Series',
-        Icon: MdLocalMovies,
-        ActiveIcon: MdLocalMovies,
-    },
-    {
-        href: 'tv shows',
-        title: 'Tv shows',
-        Icon: RiMovie2Line,
-        ActiveIcon: RiMovie2Fill,
-    },
-]
-
-const sidebarData2 = [
-    {
-        href: 'recent',
-        title: 'Recent',
-        Icon: AiOutlineClockCircle,
-        ActiveIcon: AiFillClockCircle,
-    },
-    {
-        href: 'bookmarked',
-        title: 'Bookmarked',
-        Icon: BsBookmarkStar,
-        ActiveIcon: BsBookmarkStarFill,
-    },
-    {
-        href: 'top rated',
-        title: 'Top rated',
-        Icon: AiOutlineStar,
-        ActiveIcon: AiFillStar,
-    },
-    {
-        href: 'downloaded',
-        title: 'Downloaded',
-        Icon: BiDownload,
-        ActiveIcon: BiDownload,
-    },
-]
-
-const sidebarData3 = [
-    {
-        href: 'settings',
-        title: 'Settings',
-        Icon: AiOutlineSetting,
-        ActiveIcon: AiTwotoneSetting,
-    },
-    {
-        href: 'help',
-        title: 'Help',
-        Icon: IoHelpCircleOutline,
-        ActiveIcon: IoHelpCircleSharp,
-    },
-]
-
 function Sidebar({ className }: Props) {
+    const [isCollapsed, setIsCollapsed] = useState(false)
+
+    // toggle sidebar-collapse
+    const toggleSidebar = () => {
+        setIsCollapsed(!isCollapsed)
+    }
+
     return (
-        <aside className={clsx('sidebar', className)}>
-            <SidebarMenu title='Menu' data={sidebarData1} />
-            <div className='pc:hidden'>
-                <SidebarMenu data={navbarData} />
+        <div className='flex relative'>
+            <aside
+                className={clsx('sidebar', className, isCollapsed ? 'w-[70px]' : 'w-[300px] p-4')}
+            >
+                <SidebarMenu title='Menu' data={sidebarData1} isCollapsed={isCollapsed} />
+                <div className='pc:hidden'>
+                    <SidebarMenu data={navbarData} />
+                </div>
+                <SidebarMenu title='Library' data={sidebarData2} isCollapsed={isCollapsed} />
+                <div className='mb-14'>
+                    <SidebarMenu data={sidebarData3} isCollapsed={isCollapsed} />
+                </div>
+            </aside>
+
+            {/* click sidebar collapse */}
+            <div
+                className='flex items-center justify-center rounded-full absolute top-8 -right-2 w-6 bg-blue-500 hover:scale-105 
+            h-6 cursor-pointer transition-all duration-300 ease-in-out z-10'
+                onClick={toggleSidebar}
+            >
+                {isCollapsed ? (
+                    <AiOutlineRight className='text-sm text-white' />
+                ) : (
+                    <AiOutlineLeft className='text-sm text-white' />
+                )}
             </div>
-            <SidebarMenu title='Library' data={sidebarData2} />
-            <div className='mb-14'>
-                <SidebarMenu data={sidebarData3} />
-            </div>
-        </aside>
+        </div>
     )
 }
