+++ src/components/Thumbnail2.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx'
-import { memo } from 'react'
+import { memo, useState } from 'react'
 import { BsCheck } from 'react-icons/bs'
 import { MdAdd } from 'react-icons/md'
 import { Link } from 'react-router-dom'
